✨ feat:"몬스터 HP 구현"

+++ Assets/02. Scripts/Heroes/Data/HeroData.cs
@@ -17,7 +17,7 @@ namespace LuckyDefense.Heroes.Data
         public HeroRace RaceType;
 
         [Header("ÀüÅõ")]
-        public float AttackPower;
+        public int AttackPower;
         public float AttackSpeed;
         public float Range;
